requirements.txt
openai
streamlit
langchain
faiss-cpu
python-dotenv
unstructured
pymupdf
